Change highlight code structure
Moved into multiple separate files (could become a submodule)
