+++ salve_ipc/server_functions/misc.py
@@ -1,26 +1,6 @@
 from functools import cache
 from unicodedata import category
 
-Token = tuple[tuple[int, int], int, str]
-
-generic_tokens: list[str] = [
-    "Whitespace",
-    "Text",
-    "Error",
-    "Keyword",
-    "Name",
-    "String",
-    "Number",
-    "Literal",
-    "Operator",
-    "Punctuation",
-    "Comment",
-    "Generic",
-    "Link",  # Website link (Not given by pygments)
-    "Hidden_Char",  # Hidden chars (no width space kind of stuff)
-    "Definition",  # Definitions
-]
-
 
 @cache
 def is_unicode_letter(char: str) -> bool:
